+++ examples/knapsack/main.rs
@@ -208,7 +208,7 @@ fn main() {
             .with_evaluation(&problem)
             .with_selection(MaximizeSelector::new(0.85, 12).unwrap())
             .with_crossover(SinglePointCrossBreeder::new())
-            .with_mutation(RandomValueMutator::new(0.2, false, true))
+            .with_mutation(RandomValueMutator::new(0.2, false, true).unwrap())
             .with_reinsertion(ElitistReinserter::new(&problem, false, 0.85).unwrap())
             .with_initial_population(initial_population)
             .build(),